// src/app/models/Starship.ts
export interface Starship {
  name: string;
  model: string;
  starship_class: string;
  manufacturer: string;
  length: string;
  cost_in_credits: string;
  crew: string;
  passengers: string;
  max_atmosphering_speed: string;
  hyperdrive_rating: string;
  MGLT: string;
  cargo_capacity: string;
  consumables: string;
  pilots: Array<string>;
  films: Array<string>;
  url: string;
  created: string;
  edited: string;
}